% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Basis.r
\name{Basis}
\alias{Basis}
\title{Basis class}
\description{
This class defines a basis
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Basis$new()}}
\item \href{#method-mkNewWith}{\code{Basis$mkNewWith()}}
\item \href{#method-clone}{\code{Basis$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Basis$new(x, input, fun, intercept = FALSE, cen = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mkNewWith"></a>}}
\if{latex}{\out{\hypertarget{method-mkNewWith}{}}}
\subsection{Method \code{mkNewWith()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Basis$mkNewWith(input)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Basis$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
