% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Crossbasis.r
\name{computeCrossbasis}
\alias{computeCrossbasis}
\title{Compute the crossbasis by first lagging the variable (basisvar) and then multiplying by the lag.}
\usage{
computeCrossbasis(cb)
}
\description{
Compute the crossbasis by first lagging the variable (basisvar) and then multiplying by the lag.
}
