% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Crosspred.r
\name{Crosspred}
\alias{Crosspred}
\title{Crosspred class}
\description{
This class defines and computes a prediction for Crosspred
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Crosspred$new()}}
\item \href{#method-plotExposure}{\code{Crosspred$plotExposure()}}
\item \href{#method-plotLagAt}{\code{Crosspred$plotLagAt()}}
\item \href{#method-clone}{\code{Crosspred$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Crosspred$new(
  crossbasis,
  model,
  at,
  coef = NULL,
  vcov = NULL,
  cen = NULL,
  ci.level = 0.95,
  cumul = FALSE,
  model.link = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotExposure"></a>}}
\if{latex}{\out{\hypertarget{method-plotExposure}{}}}
\subsection{Method \code{plotExposure()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Crosspred$plotExposure(filepath = NULL, title = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotLagAt"></a>}}
\if{latex}{\out{\hypertarget{method-plotLagAt}{}}}
\subsection{Method \code{plotLagAt()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Crosspred$plotLagAt(
  at,
  filepath = NULL,
  title = NULL,
  xDist = 0.1,
  colours = colourPalette,
  legendPos = c(0.5, 0.1)
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Crosspred$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
