% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CrosspredInteraction.r
\name{CrosspredInteraction}
\alias{CrosspredInteraction}
\title{Crosspred class}
\description{
This class defines and computes a prediction for Crosspred
}
\section{Super class}{
\code{\link[dlnm.interaction:Crosspred]{dlnm.interaction::Crosspred}} -> \code{CrosspredInteraction}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{CrosspredInteraction$new()}}
\item \href{#method-plotExposure}{\code{CrosspredInteraction$plotExposure()}}
\item \href{#method-plotLagAt}{\code{CrosspredInteraction$plotLagAt()}}
\item \href{#method-clone}{\code{CrosspredInteraction$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CrosspredInteraction$new(
  crossbasisExposure,
  crossbasisInteraction,
  model,
  at,
  coefInteraction = NULL,
  coefExposure = NULL,
  vcovInteraction = NULL,
  vcovExposure = NULL,
  vcovIntersection = NULL,
  cen = NULL,
  ci.level = 0.95,
  cumul = FALSE,
  model.link = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotExposure"></a>}}
\if{latex}{\out{\hypertarget{method-plotExposure}{}}}
\subsection{Method \code{plotExposure()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CrosspredInteraction$plotExposure(filepath = NULL, title = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotLagAt"></a>}}
\if{latex}{\out{\hypertarget{method-plotLagAt}{}}}
\subsection{Method \code{plotLagAt()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CrosspredInteraction$plotLagAt(
  at,
  filepath = NULL,
  title = NULL,
  xDist = 0.1,
  colours = colourPalette,
  legendPos = c(0.5, 0.1)
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CrosspredInteraction$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
