% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logknots.r
\name{logknots}
\alias{logknots}
\title{Logknots}
\usage{
logknots(lag, nk)
}
\arguments{
\item{lag}{numeric/vector  maximum lag, lag range or vector of lags}

\item{nk}{numeric         number of knots (must be >0)}
}
\value{
\preformatted{vector
}
}
\description{
Logknots
}
