% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Crossbasis.r
\name{mkLaggedMatrix}
\alias{mkLaggedMatrix}
\title{Lag vector \code{x} by a given vector of \code{lags}}
\usage{
mkLaggedMatrix(x, lags)
}
\description{
Lag vector \code{x} by a given vector of \code{lags}
}
