% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ns.r
\name{Ns}
\alias{Ns}
\title{Linear Basis}
\description{
Linear Basis

Linear Basis
}
\section{Super class}{
\code{\link[dlnm.interaction:Basis]{dlnm.interaction::Basis}} -> \code{Ns}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Ns$new()}}
\item \href{#method-mkNewWith}{\code{Ns$mkNewWith()}}
\item \href{#method-clone}{\code{Ns$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ns$new(x, knots, intercept = FALSE, Boundary.knots = range(x))}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mkNewWith"></a>}}
\if{latex}{\out{\hypertarget{method-mkNewWith}{}}}
\subsection{Method \code{mkNewWith()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ns$mkNewWith(x)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ns$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
