% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_exposure.r
\name{plotExposure}
\alias{plotExposure}
\title{Retrieve GGPlot object for the Relative Risk (RR) dependend on the Exposure of a Crosspred object.}
\usage{
plotExposure(
  cp,
  filepath = NULL,
  title = NULL,
  colours = colourPalette,
  legendPos = c(0.5, 0.1)
)
}
\arguments{
\item{cp}{Crosspred object to plot.}

\item{filepath}{Filepath to save the plot.}

\item{title}{Title of plot.}
}
\value{
object     GGPlot object. Execute, i.e. do not save it in a variable, to plot.
}
\description{
Retrieve GGPlot object for the Relative Risk (RR) dependend on the Exposure of a Crosspred object.
}
