% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_lag.r
\name{plotLagAt}
\alias{plotLagAt}
\title{Compute GGPlot object for the Relative Risk (RR) dependend on the Lag of a Crosspred object.}
\usage{
plotLagAt(
  cp,
  filepath = NULL,
  title = NULL,
  xDist = 0.1,
  colours = colourPalette,
  legendPos = c(0.5, 0.1)
)
}
\arguments{
\item{cp}{Crosspred object to plot.}

\item{filepath}{Filepath to save the plot.}

\item{title}{Title of plot.}
}
\value{
object     GGPlot object. Execute, i.e. do not save it in a variable, to plot.
}
\description{
Compute GGPlot object for the Relative Risk (RR) dependend on the Lag of a Crosspred object.
}
