% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{savePlot}
\alias{savePlot}
\title{Function used for saving the plots.}
\usage{
savePlot(
  filename,
  plot,
  plotSaveFormat = "pdf",
  plotDpi = 300,
  cols = 1,
  rows = 1,
  plotWidth = 10,
  plotHeight = 3
)
}
\arguments{
\item{name}{string.}
}
\description{
Function used for saving the plots.
}
